Mock Deno.exit() in the test to throw an error
This is required as Deno.exit(1) after `error()` is making `test.execute()` exit without possibility to cath the error and log the error to be verified.

Our test logic does not even fail on this Deno.exit() happening. Tests continue and nor OK or FAILED is thrown.

Our usual pattern to throw an error when using error(), but it may not apply in this case for checking version requirement in quarto.ts. This is why mocking is chosen here.