Add QUARTO_KNITR_RSCRIPT_ARGS environment variable
to pass some flags to Rscript run by Quarto for knitr engine. This can
be useful in specific situation like e.g `--max-connections=258` available in R 4.4, or `--vanilla`

Args should be passed a comma separated list of flags.

This is for expert use and there is not check done on the argument being support by RScript or not.