Return modified meta object from init filter
If the metadata object is modified in a filter, then it should be
returned, as otherwise the changes might be lost.

In the given case, the return wasn't necessary, because filters that act
only on `Meta` objects are handled specially, in that the filter
function is called directly, without doing an actual traversal via
`walk`. However, this can lead to unexpected bugs; explicitly returning
the modified Meta object is cleaner and more robust.