Update `QuartoNotebookRunner.jl` to `0.13.1`
Updates the backend to the latest minor release.

This includes support for running `{python}` code cells (in addition to the already supported `{julia}` and `{r}`). It uses https://github.com/JuliaPy/PythonCall.jl to manage the `python` process for this, and allows passing values via `$` interpolation between all three languages.

The announcement post can be found at https://discourse.julialang.org/t/ann-native-julia-engine-for-quarto-using-quartonotebookrunner-jl/112753/56, and includes an example of using this new feature.