Revert "gha - opt out Julia Cache"
This reverts commit 210961e34e32fe2a402f58efb9a17d24b4f9890a.

Julia cache is required to speed up each job. However, the caching mechanism is still a problem as it will throw a warning for each other job than the one which managed to save with action v4

Saving a cache for each buckets would be too heavy in total storage (we are limited to 10Go).