latex, template - pandoc now uses partials for LaTeX and split beamer in its own template
This is a breaking change for users who have custom templates. The new templates are more modular and easier to customize, but they now mix Pandoc's partial and Quarto's partials.

Changes:

- Add an new template logic dedicated to beamer as Pandoc has split beamer in its own template. Aim is to make it easy to update the beamer template in the future.
- Add a new LaTeX template that is based on the default template, but it is split into several partials that are provided by Pandoc and by Quarto.
- Adapt the update script to use the new partials and to update the beamer template separately.
- Add two Quarto partials that are used in the new LaTeX template: `babel-lang.tex` and `biblio-config.tex`.

The new organization possibly could break custom format that tweak the `pandoc.tex` partials as it now contains part of the default `common.latex` partial from Pandoc. This means that if you have a custom template that uses `pandoc.tex` you will need to update it to use the new partials.