feat: allow labels for multiple top-level figures
Emit a warning if a user combines multiple top-level figures with
labels.

Use the same mechanism as for sub-figures to number them.
This will result in the original label without a number no existing, as
it would no make any sense to have it. Without this change the label
points to the last figure.