Use jsdeliver cdn for jupyter widgets dependencies
This PR replaces the use of cdnjs.cloudflare.com and unpkg.com with cdn.jsdelivr.net for jupyter widgets. The motivation is to bring more uniformity to JS CDN's for people defining content security policies for quarto websites. Since we already use jsdelivr for MathJax, Katex, and Algolia this brings Juptyer Widgets in line with those uses.