Correctly normalize to forward slash the path when creating tar.gz for Connect deploys
std/tar will create folders based on file paths used, but it requires using / and not \\ in path. This is a problem on Windows, where the path separator is \\. This commit normalize the path to use / instead of \\ when creating the tar.gz file for Connect deploys.

(cherry picked from commit ffb74591f7484c247b59c51b0e0dc972a834f1df)