minimal: true should disable brand, not crash
fixes #13383

v1.7 introduced this code implementing the light, dark, light
stylesheet trick for syntax highlighting css.

it should not try to do this if there are no stylesheets!

we did not see the bug trigger until v1.8 which treats brand
data as possibly implementing both light and dark
decided when we see theme.